# OpenApiGenerator/src/ProductInventoryApi/Controllers/InventoryApi.cs
/*
 * Product Inventory API
 *
 * API for managing product inventory
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using ProductInventoryApi.Attributes;
using ProductInventoryApi.Models;

namespace ProductInventoryApi.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class InventoryApiController : ControllerBase
    { 
        /// <summary>
        /// Get inventory status
        /// </summary>
        /// <response code="200">Inventory status</response>
        [HttpGet]
        [Route("/api/v1/inventory")]
        [ValidateModelState]
        [SwaggerOperation("GetInventory")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<InventoryItem>), description: "Inventory status")]
        public abstract IActionResult GetInventory();

        /// <summary>
        /// Get inventory for a product
        /// </summary>
        /// <param name="productId">ID of the product</param>
        /// <response code="200">Product inventory</response>
        /// <response code="404">Product not found</response>
        [HttpGet]
        [Route("/api/v1/inventory/{productId}")]
        [ValidateModelState]
        [SwaggerOperation("GetProductInventory")]
        [SwaggerResponse(statusCode: 200, type: typeof(InventoryItem), description: "Product inventory")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "Product not found")]
        public abstract IActionResult GetProductInventory([FromRoute (Name = "productId")][Required]string productId);

        /// <summary>
        /// Update inventory for a product
        /// </summary>
        /// <param name="productId">ID of the product</param>
        /// <param name="inventoryUpdate"></param>
        /// <response code="200">Inventory updated</response>
        /// <response code="400">Invalid input</response>
        /// <response code="404">Product not found</response>
        [HttpPut]
        [Route("/api/v1/inventory/{productId}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("UpdateProductInventory")]
        [SwaggerResponse(statusCode: 200, type: typeof(InventoryItem), description: "Inventory updated")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Invalid input")]
        [SwaggerResponse(statusCode: 404, type: typeof(Error), description: "Product not found")]
        public abstract IActionResult UpdateProductInventory([FromRoute (Name = "productId")][Required]string productId, [FromBody]InventoryUpdate inventoryUpdate);
    }
}